+++ src/message.ts
@@ -262,6 +262,7 @@ export class StunMessage {
     // check if integrity w/ fingerprint
     return generateIntegrityWithFingerprint(
       this.toBuffer(),
+      this.header.length,
       integrityKey,
     ).equals(messageIntegrityAttr.value);
   }
