+++ src/utils/stun.ts
@@ -1,3 +1,7 @@
+import * as crc32 from 'buffer-crc32';
+import { bufferXor } from './buffer';
+import { generateHmacSha1Digest } from './crypto';
+
 export function isStunMessage($buffer: Buffer): boolean {
   // the first 2bit must be 0 in first 1byte
   const first8bit = numberToBinaryStringArray($buffer[0], 8);
@@ -115,4 +119,27 @@ export function writeAttrBuffer(type: number, $value: Buffer): Buffer {
   const $padding = Buffer.alloc(paddingByte);
 
   return Buffer.concat([$type, $length, $value, $padding]);
+}
+
+export function generateFingerprint($msg: Buffer): Buffer {
+  // without FINGERPRINT: 8byte(header: 4byte + value: 4byte(32bit))
+  const $crc32 = crc32($msg.slice(0, -8));
+  return bufferXor($crc32, Buffer.from('5354554e', 'hex'));
+}
+
+export function generateIntegrity($msg: Buffer, integrityKey: string): Buffer {
+  // without MESSAGE-INTEGRITY: 24byte(header: 4byte + value: 20byte(sha1))
+  return generateHmacSha1Digest(integrityKey, $msg.slice(0, -24));
+}
+
+export function generateIntegrityWithFingerprint(
+  $msg: Buffer,
+  msgLen: number,
+  integrityKey: string,
+): Buffer {
+  // modify header length to ignore FINGERPRINT(8byte)
+  $msg.writeUInt16BE(msgLen - 8, 2);
+  // without 32byte MESSAGE-INTEGRITY(24byte)
+  // + FINGERPRINT: 8byte = (header: 4byte + value: 4byte)
+  return generateHmacSha1Digest(integrityKey, $msg.slice(0, -32));
 }
